it compiles

+++ src/lib.rs
@@ -13,7 +13,7 @@ pub const MTU: usize = 1420;
 const MAX_UDP_SIZE: usize = (1 << 16) - 1;
 pub const CHANNEL_BUFFER_SIZE: usize = MTU + 512; // Buffered channels
 pub const ENCRYPTION_OVERHEAD: usize = 28; // 12 nonce + 16 auth tag
-pub const KEEPALIVE_INTERVAL: usize = 26;
+pub const KEEPALIVE_INTERVAL: u64 = 26;
 
 // types
 #[derive(serde::Serialize, serde::Deserialize, PartialEq, Debug, Clone)]
